Add login UI to sync (needs work)
There are still lifecycle issues when the credentials sync completes. It
leaves you in a weird state, and gives a bad message when you fail to
authenticate. This is probably because I'm applying the
VerifyServerSettings state machine in a way it wasn't intended.
Furthermore, we need to validate the database and tell the user to sync
before letting them change their credentials