Fix: abort of sync when writing device status record.
The device status record can only be sent if we are authenticating successfully.
If we are an anonymous user or the server is not an ODK server, does not accept
our appName, reports an authentication error, or network transport error, then don't send the device status record. And, if we do send it, catch all exceptions and swallow them (vs. just IOException).