fix: always detect changes to app.properties after sync
After the app-level files have been sync'd, successfully or not, increment the properties file revision tracking number so that client apps will re-read their properties from the possibly-revised disk files.  This is independent of the re-initialization logic fix, which affects the scanning of the folders for change. If the revision number is not incremented, the client apps will not detect app.properties changes.