fix: potential inconsistencies and crashes
Two fixes for potential inconsistencies and crashes due to shared access to data structures. 
(1) ODKDatabaseImplUtils -- the user roles are expanded into a List<String> and cached. This fix changes getRolesArray() to an instance method and adds a cacheGuard object which is used to synchronize access to the roles list and roles array. Without this change, there is a possibility that the returned list of roles might not correspond to the roles that are active (detected on website).
(2) The OperationLog maintains a circular buffer of the last N operations and their outcomes. There were issues with proper access to this buffer only within a synchronized block on the mOperations buffer array.  The primary change is to the getOperationLocked() method (line 320). This now operates within that synchronized block. Other changes are to private methods, which are now passed the OperationLogEntry rather than the cookie used to retrieve the entry.