remove return values for insert and replace APIs; add factory.
These are changes needed to maximize code reuse on website:
(1) SQLiteConnection split into SQLiteConnectBase abstract class and the implementation class. 
(2) SQLiteConnectionFactory class added. Creates a new SQLiteConnection.
(3) Remove the return values for insert and replace. These are unused except in the FormsProvider, where the value is used to invalidate any cursors that are open against that newly created row -- which are necessarily none -- so removal of that code has no impact.